ml-service/requirements.txt
fastapi
uvicorn[standard]

pydantic
python-multipart

pymongo
motor

numpy
pandas
scikit-learn
joblib

opencv-python
pytesseract
pdf2image
Pillow

openai

python-dotenv

tqdm